+++ custom/justech_l10n_do_ncf/wizards/__init__.py
@@ -1,1 +1,3 @@
 from . import fiscal_diagnostic_wizard
+from . import ncf_migration_wizard
+from . import ncf_reconcile_wizard